[geometrysampler] remove point argument from () operator
Geometry samplers can now be called to return a random geometry by
simply calling the () operator, without the need to externally compute
the point around which the geometry is to be created. This means that
the point sampler is now contained in the geometry sampler class. The
only currently available implementation (DiskSampler) has been modified
accordingly.