[intersection] remove disk-shell algorithm
The implementation did not detect multiple face intersections and also
did not find touching points of the shell inside the disk. Since
for the constraints checks this is obsolete as the can be checked for
all faces of a shape, we remove this implementation here and might
reintroduce it at one point in the future.