[geometry] make all geometries default-constructible
This could be usefol, for example, in the case that one passes
geometries as references to a function in which they are actually
created. At the point of creating these objects in order to pass them to
a function one might not have enough data/information to actually
instantiate them.