[geometry] implement bounding box
For now, this is just an alias for the Box class, but in case the Box
class is extended at one point to be more general, i.e. to allow for
non-axis-aligned boxes, it wouldn't serve as an implementation for
bounding boxes anymore. To this end, we now introduce this alias which
allows us to mark interfaces that require bounding boxes to be passed to
them already at this point.