[intersect] check if bboxes intersect before calling algorithm
This leads to a substantial speed up in the intersection time
measurement test for the empty intersections.
TODO: We could introduce a compile-time option to allow users to
deactivate this upon configuration. Because, in case one knows that
almost all entities that will be created will intersect, the query is
obsolete. However, for sparse distributions of entities, this will
speedup constraints evaluations significantly.