[sampling][quad] use polygon sampler internally
This commit deprecates the old implementation that received a
distribution for the edge length, in favour of two distributions for the
characteristic dimensions in dip & strike direction. Internally, the new
polygon sampler is used, and when construction in the old way is
intended, a deprecation warning is thrown. However, this only works with
the default traits. Compilation will fail if custom traits were used...